WhatsappDataManager:   Fixed https://github.com/Auties00/WhatsappWeb4j/issues/36   Fixed message processing from node   Refactored code for better readability   Added handy addChat method   Added handler for server message that checks type instead of assuming that it's a revoke message type WhatsappAPI:   Fixed markChat method, the issue was caused by the first issue mentioned in WhatsappDataManager

+++ src/main/java/it/auties/whatsapp4j/protobuf/message/server/ProtocolMessage.java
@@ -41,7 +41,7 @@ public final class ProtocolMessage implements ServerMessage {
    * This property is defined only if {@link ProtocolMessage#type} == {@link ProtocolMessageType#EPHEMERAL_SETTING} || @link ProtocolMessageType#EPHEMERAL_SYNC_RESPONSE}.
    */
   @JsonProperty(value = "4")
-  private int ephemeralExpiration;
+  private long ephemeralExpiration;
 
   /**
    * The type of this server message
